add: more user api

+++ src/utils.php
@@ -15,9 +15,7 @@ function base64url_encode($data)
 
 function base64url_decode($data)
 {
-    return base64_decode(
-        str_pad(strtr($data, "-_", "+/"), strlen($data) % 4, "=", STR_PAD_RIGHT)
-    );
+    return base64_decode(str_pad(strtr($data, "-_", "+/"), strlen($data) % 4, "=", STR_PAD_RIGHT));
 }
 
 function uploadFiles(array $files, string $dirKey, array $options = []): array
@@ -28,22 +26,14 @@ function uploadFiles(array $files, string $dirKey, array $options = []): array
         throw new Error("Direktori dengan key '$dirKey' tidak ditemukan.", 500);
     }
 
-    if (empty($files)) {
-        return [];
-    }
+    if (empty($files)) return [];
 
     // Options
-    $prefix =
-        isset($options["prefix"]) && checkType($options["prefix"], "string")
-        ? $options["prefix"]
-        : "";
-    $suffix =
-        isset($options["suffix"]) && checkType($options["suffix"], "string")
-        ? $options["suffix"]
-        : "";
-    $date =
-        isset($options["date"]) && checkType($options["date"], "boolean")
-        ? $options["date"]
+    $prefix = $options["prefix"] ?? "";
+    $suffix = $options["suffix"] ?? "";
+    $withDate =
+        isset($options["withDate"]) && checkType($options["withDate"], "boolean")
+        ? $options["withDate"]
         : false;
 
     $savedPaths = [];
@@ -53,7 +43,7 @@ function uploadFiles(array $files, string $dirKey, array $options = []): array
         $extension = pathinfo($file["name"], PATHINFO_EXTENSION);
 
         $dt = new DateTime();
-        $timestamp = $date ? "_" . $dt->format("YmdHisv") : "";
+        $timestamp = $withDate ? "_" . $dt->format("YmdHisv") : "";
 
         $filename = $prefix . $originalName . $timestamp . $suffix;
         $filename .= $extension ? "." . $extension : "";
@@ -73,9 +63,7 @@ function uploadFiles(array $files, string $dirKey, array $options = []): array
 function removeFiles(array $filePaths)
 {
     foreach ($filePaths as $filePath) {
-        if (file_exists($filePath)) {
-            unlink($filePath);
-        }
+        if (file_exists($filePath)) unlink($filePath);
     }
 }
 
@@ -84,19 +72,14 @@ function action(
     callable $callback,
     array $response = []
 ) {
-    if (!checkMethod($methods)) {
-        return;
-    }
-
+    if (!checkMethod($methods)) return;
     global $db;
 
     try {
         call_user_func_array($callback, [$db]);
     } catch (Throwable $th) {
         $onError = $response["onError"] ?? null;
-        if ($onError && is_callable($onError)) {
-            call_user_func_array($onError, [$th]);
-        }
+        if ($onError && is_callable($onError)) call_user_func_array($onError, [$th]);
         responseError($th);
     }
 }
